Increased testing coverage to > 80%
