+++ .github/workflows/ci_build_tests.yml
@@ -16,7 +16,6 @@ jobs:
   build:
 
     runs-on: ubuntu-latest
-
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python ${{ matrix.python-version }}
@@ -37,8 +36,6 @@ jobs:
     - name: Run Unit Tests
       run: pytest ./tests/unit_tests --cov=. --cov-append --cov-config .coveragerc        
     - name: Run Integration Tests
-      run: pytest ./tests/integration_tests --cov=. --cov-append --cov-config .coveragerc             
-    - name: Run dataset tests
-      run: pytest ./tests/integration_tests/test_dataset_ipfs.py --cov=. --cov-append --cov-config .coveragerc        
+      run: pytest ./tests/integration_tests --cov=. --cov-append --cov-config .coveragerc                  
     - name: Upload Coverage to Codecov
       uses: codecov/codecov-action@v1
