+++ bittensor/_wandb/__init__.py
@@ -36,7 +36,6 @@ class wandb:
             tags: tuple = None,
             run_group: str = None,
             directory: str = None,
-            offline: bool = None,
             cold_pubkey: str = None,
             hot_pubkey: str = None,
             root_dir: str = None
@@ -60,6 +59,7 @@ class wandb:
         os.environ["WANDB_PROJECT"] = config.wandb.project if config.wandb.project != 'default' else str(cold_pubkey)[:8]
         os.environ["WANDB_TAGS"] = config.wandb.tags 
         os.environ["WANDB_RUN_GROUP"] = config.wandb.run_group if config.wandb.run_group != 'default' else str(hot_pubkey)[:8]
+        import pdb; pdb.set_trace()
         os.environ["WANDB_DIR"] = config.wandb.directory if config.wandb.directory != 'default' else root_dir
 
         wb.init(config = config, config_exclude_keys = ['wandb'])
