+++ tests/integration_tests/subtensor/v2.0.0/client.py
@@ -328,3 +328,33 @@ def test_set_weights_success(setup_chain):
     assert result_uids == w_uids.tolist()
     assert result_vals == w_vals.tolist()
 
+def test_solve_for_difficulty_fast( setup_chain ):
+    subtensor, port = setup_subtensor(setup_chain)
+    nonce, block_number, block_hash, difficulty, seal = bittensor.utils.solve_for_difficulty_fast(subtensor)
+
+    assert difficulty == 10000
+
+def test_indexed_values_to_dataframe( setup_chain ):
+    subtensor, port = setup_subtensor(setup_chain)
+    wallet = generate_wallet()
+
+    wallet.register(subtensor=subtensor)
+    nn = subtensor.neuron_for_pubkey(wallet.hotkey.ss58_address)
+    metagraph = bittensor.metagraph(subtensor=subtensor)
+    metagraph.sync()
+    idx_df = bittensor.utils.indexed_values_to_dataframe( prefix = 'w_i_{}'.format(nn.uid), index = [nn.uid], values = metagraph.W[:, nn.uid] )
+    assert idx_df.values[0][0] == 1.0
+
+    idx_df = bittensor.utils.indexed_values_to_dataframe( prefix = nn.uid, index = [nn.uid], values = metagraph.W[:, nn.uid] )
+    assert idx_df.values[0][0] == 1.0
+
+    idx_df = bittensor.utils.indexed_values_to_dataframe( prefix = nn.uid, index = torch.LongTensor([nn.uid]), values = metagraph.W[:, nn.uid] )
+    assert idx_df.values[0][0] == 1.0
+
+    # Need to check for errors 
+    with pytest.raises(ValueError):
+        idx_df = bittensor.utils.indexed_values_to_dataframe( prefix = b'w_i', index = [nn.uid], values = metagraph.W[:, nn.uid] )
+    with pytest.raises(ValueError):
+        idx_df = bittensor.utils.indexed_values_to_dataframe( prefix = 'w_i_{}'.format(nn.uid), index = nn.uid, values = metagraph.W[:, nn.uid] )
+    with pytest.raises(ValueError):
+        idx_df = bittensor.utils.indexed_values_to_dataframe( prefix = 'w_i_{}'.format(nn.uid), index = [nn.uid], values = nn.uid)
